Use HUGO, port all documentation to pages, host pages in GitHub Pages
The whole documentation was first used in GitHub Wiki. While pretty useful for smaller projects, this made it very cumbersome for others to contribute as it was its own hidden repository that isn't accessible for PRs.
This creates a full blown HUGO website with the documentation. All Wiki pages have been ported over and fit for the new structure.
This will make it very easy for anybody to contribute and add in changes. Also for normal users, it's just a better experience to browse than through the GitHub Wiki.
Additionally, there's pages for contribution and community info to present the other projects better.
All this is automatically compiled with GitHub Actions and hosted on GitHub pages.